[APDV-111] Moved unit converter's mapper to separate package
