+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/unit_converter/UpdateUnitConverterImpl.java
@@ -2,8 +2,8 @@ package pl.edu.agh.apdvbackend.use_cases.unit_converter;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.edu.agh.apdvbackend.mappers.unit.AddUnitConverterRequestBodyMapper;
-import pl.edu.agh.apdvbackend.mappers.unit.UnitConverterResponseBodyMapper;
+import pl.edu.agh.apdvbackend.mappers.unit_converter.AddUnitConverterRequestBodyMapper;
+import pl.edu.agh.apdvbackend.mappers.unit_converter.UnitConverterResponseBodyMapper;
 import pl.edu.agh.apdvbackend.models.body_models.unit_converter.AddUnitConverterRequestBody;
 import pl.edu.agh.apdvbackend.models.body_models.unit_converter.UnitConverterResponseBody;
 import pl.edu.agh.apdvbackend.repositories.UnitConverterRepository;
